{"schema":"decodex/commit/1","summary":"Split MCP tests","authority":"XY-1115"}
